+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - N/A
+### Added
+- PSR-6 compliant Cache
 
 ## [0.0.4] - 2017-04-02
 ### Added